newbackend/requirements.txt
flask==2.2.2
Werkzeug==2.2.2
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
Pillow
cffi

